+++ CHANGELOG.md
@@ -26,6 +26,8 @@ omnibus-gitlab repository.
 - Update Prometheus redis_exporter to v0.26.0 !2981
 - Support smartcard auth against LDAP server !3006
 - Update Mattermost to 5.7.1
+- Update Prometheus alerting rules !3011
+- Allow external diffs for merge requests to be configured
 
 11.7.0